+++ client/src/main/java/replicant/ReplicantConfig.java
@@ -26,7 +26,7 @@ final class ReplicantConfig
   private static boolean ENABLE_SPIES =
     "true".equals( System.getProperty( "replicant.enable_spies", PRODUCTION_MODE ? "false" : "true" ) );
   /**
-   * Valid values are: "none", "basic", "jul" (java.util.logging) and "proxy" (for testing)
+   * Valid values are: "none", "console" and "proxy" (for testing)
    */
   private static final String LOGGER_TYPE =
     System.getProperty( "arez.logger", PRODUCTION_MODE ? "basic" : "proxy" );