+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Unreleased
 
-* Upgrade the `org.realityforge.braincheck` artifact to version `1.24.0`.
+* Upgrade the `org.realityforge.braincheck` artifact to version `1.25.0`.
 
 ### [v6.46](https://github.com/replicant4j/replicant/tree/v6.46) (2019-09-16) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.45...v6.46)