+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-### Unreleased
+### [v6.48](https://github.com/replicant4j/replicant/tree/v6.48) (2019-10-18) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.47...v6.48)
 
 * Upgrade the `org.realityforge.arez` artifacts to version `0.151`.
 * Suppress or eliminate rawtypes warnings and turn on linting to avoid their return.